add data_buffer & task_pool.

+++ sys/db/data_buffer_test.go
@@ -38,7 +38,7 @@ func TestAddData(t *testing.T) {
 	})
 
 	metaData := "testMetaData"
-	dataID := uint64(1)
+	dataID := uint(1)
 
 	success := dbBox.AddData(dataID, metaData)
 	if !success {
@@ -64,7 +64,7 @@ func TestDeleteData(t *testing.T) {
 	})
 
 	metaData := "testMetaData"
-	dataID := uint64(1)
+	dataID := uint(1)
 
 	dbBox.AddData(dataID, metaData)
 	dbBox.DeleteData(dataID)
